+++ packages/ra-no-code/src/ApplicationsDashboard/applicationStorage.ts
@@ -1,0 +1,18 @@
+export const loadApplicationsFromStorage = () => {
+    const storedValue = window.localStorage.getItem(
+        '@@ra-no-code/applications'
+    );
+
+    if (storedValue) {
+        return JSON.parse(storedValue);
+    }
+
+    return [];
+};
+
+export const storeApplicationsInStorage = applications => {
+    window.localStorage.setItem(
+        '@@ra-no-code/applications',
+        JSON.stringify(applications)
+    );
+};
